requirements.txt
streamlit
pandas
pytesseract
Pillow
